requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==4.2.16
django-debug-toolbar==4.4.6
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
model-bakery==1.20.0
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
typing-extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
